// ui-inbox/src/inbox/types.ts
export interface ICarouselButton {
  type: string;
  title: string;
  text: string;
  payload: string;
  url?: string;
}
export interface ICarouselItem {
  title: string;
  picture?: string;
  file?: string;
  video?: string;
  subtitle?: string;
  footer?: string;
  buttons?: ICarouselButton[];
}